Version 3.29.82 (based on bleeding_edge revision r24129)
Fix escaped index JSON parsing (Chromium issue 416449).

Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@24136 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/arm64/simulator-arm64.cc
@@ -1855,9 +1855,12 @@ void Simulator::LoadStoreWriteBack(unsigned addr_reg,
 void Simulator::CheckMemoryAccess(uintptr_t address, uintptr_t stack) {
   if ((address >= stack_limit_) && (address < stack)) {
     fprintf(stream_, "ACCESS BELOW STACK POINTER:\n");
-    fprintf(stream_, "  sp is here:          0x%016" PRIx64 "\n", stack);
-    fprintf(stream_, "  access was here:     0x%016" PRIx64 "\n", address);
-    fprintf(stream_, "  stack limit is here: 0x%016" PRIx64 "\n", stack_limit_);
+    fprintf(stream_, "  sp is here:          0x%016" PRIx64 "\n",
+            static_cast<uint64_t>(stack));
+    fprintf(stream_, "  access was here:     0x%016" PRIx64 "\n",
+            static_cast<uint64_t>(address));
+    fprintf(stream_, "  stack limit is here: 0x%016" PRIx64 "\n",
+            static_cast<uint64_t>(stack_limit_));
     fprintf(stream_, "\n");
     FATAL("ACCESS BELOW STACK POINTER");
   }
